requirements.txt
fastapi
uvicorn
python-kasa
tzdata